feat: Add support for larger domains
This commit lets the programmer use larger domains than NX=NY=NZ=1024

+++ include/constants.h
@@ -26,7 +26,7 @@
 
 #define NX DIM
 #define NY DIM
-#define NZ DIM
+#define NZ ((unsigned long) DIM)
 #define SIZE             (NX*NY*NZ)
 #define OFFSET           (SIZE/NGPUS)
 #define BYTES            (SIZE*sizeof(float))
